Fix some visual errors on tool breaking
Breaking particles are now head instead of missing texture, and durability bar shows empty
